creation hook useScreenWidth
