+++ client/src/components/PokeCard/PokeCard.jsx
@@ -1,4 +1,4 @@
-import { useState, useEffect } from "react";
+import { useState } from "react";
 import { useOutletContext, useRouteLoaderData } from "react-router-dom";
 import PokeName from "./PokeName";
 import Poketypes from "../PokeCardMobile/PokeTypes/PokeTypes";
@@ -6,25 +6,14 @@ import StatsPoke from "../Stats/StatsPoke";
 import CardImg from "./CardImg";
 import poketype from "../../assets/functions/poketypefunction";
 import "./PokeCard.css";
+import useScreenWidth from "../../hooks/useScreenWidth";
 
 function PokeCard() {
   const pokemonsData = useRouteLoaderData("Pokecard");
   const [pokemons] = useState(pokemonsData);
   const { random } = useOutletContext();
   const typeImgUrl = poketype(pokemons, random);
-  const [isMobile, setIsmobile] = useState(window.innerWidth < 800);
-
-  useEffect(() => {
-    const handleResize = () => {
-      setIsmobile(window.innerWidth < 800);
-    };
-
-    window.addEventListener("resize", handleResize);
-
-    return () => {
-      window.removeEventListener("resize", handleResize);
-    };
-  }, []);
+  const isMobile = useScreenWidth();
 
   return (
     <div className={`pokeCard ${typeImgUrl[0].color}`}>
