+++ client/src/components/PokedexMobile/PokedexMobile.jsx
@@ -3,6 +3,7 @@ import { useOutletContext } from "react-router-dom";
 import styles from "./PokedexMobile.module.css";
 import Pokemon from "../Pokemon/Pokemon";
 import ButtonPokedexCard from "../ButtonPokedexCard/ButtonPokedexCard";
+import useScreenWidth from "../../hooks/useScreenWidth";
 
 function PokedexMobile() {
   const { pokemons } = useOutletContext();
@@ -10,22 +11,11 @@ function PokedexMobile() {
   const { random } = useOutletContext();
   const { setIsPokedex } = useOutletContext();
   const [displayedPokemons, setDisplayedPokemons] = useState([]);
-  const [isMobile, setIsMobile] = useState(window.innerWidth < 800);
+  const isMobile = useScreenWidth();
 
   useEffect(() => {
     setIsPokedex(true);
   }, [setIsPokedex]);
-  useEffect(() => {
-    const handleResize = () => {
-      setIsMobile(window.innerWidth < 800);
-    };
-
-    window.addEventListener("resize", handleResize);
-
-    return () => {
-      window.removeEventListener("resize", handleResize);
-    };
-  }, []);
 
   return (
     <div>
@@ -34,14 +24,25 @@ function PokedexMobile() {
           isMobile ? styles.pokedexCard_mobile : styles.pokedexCard_desktop
         }
       >
-        {displayedPokemons.map((pokemon) => (
-          <Pokemon
-            key={pokemon.id}
-            pokemon={pokemon}
-            setRandom={setRandom}
-            random={random}
-          />
-        ))}
+        {isMobile === true
+          ? displayedPokemons
+              .slice(0, 6)
+              .map((pokemon) => (
+                <Pokemon
+                  key={pokemon.id}
+                  pokemon={pokemon}
+                  setRandom={setRandom}
+                  random={random}
+                />
+              ))
+          : displayedPokemons.map((pokemon) => (
+              <Pokemon
+                key={pokemon.id}
+                pokemon={pokemon}
+                setRandom={setRandom}
+                random={random}
+              />
+            ))}
       </div>
       <ButtonPokedexCard
         pokemons={pokemons}
