+++ client/src/hooks/useScreenWidth.jsx
@@ -1,0 +1,19 @@
+import { useEffect, useState } from "react";
+
+function useScreenWidth() {
+  const [isMobile, setIsMobile] = useState(window.innerWidth < 800);
+  useEffect(() => {
+    const handleResize = () => {
+      setIsMobile(window.innerWidth < 800);
+    };
+
+    window.addEventListener("resize", handleResize);
+
+    return () => {
+      window.removeEventListener("resize", handleResize);
+    };
+  }, []);
+  return isMobile;
+}
+
+export default useScreenWidth;
